+++ Control/Joint/Abilities.hs
@@ -3,4 +3,4 @@ module Control.Joint.Abilities (module Exports) where
 import Control.Joint.Abilities.Liftable as Exports
 import Control.Joint.Abilities.Adaptable as Exports
 import Control.Joint.Abilities.Transformer as Exports
-import Control.Joint.Abilities.Composition as Exports
+import Control.Joint.Abilities.Interpreted as Exports