+++ Control/Joint/Abilities/Composition.hs
@@ -1,6 +1,0 @@
-module Control.Joint.Abilities.Composition (Composition (..)) where
-
-class Composition t where
-	{-# MINIMAL run #-}
-	type Primary t a :: *
-	run :: t a -> Primary t a