+++ CHANGELOG.md
@@ -34,3 +34,4 @@
 # 0.1.6
 * Lower precedence of `:>` operator to allow write like: `t :> u := a`
 * Create supermodule to have only one line of import in projects
+* Rename `Composition` class to `Interpreted`